requirements.txt
# FastAPI and server
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
requests==2.31.0

# Face recognition
opencv-python==4.8.1.78
mtcnn==0.1.1
onnxruntime==1.16.3
insightface==0.7.3

# Database
cassandra-driver==3.28.0

# Image processing
Pillow==10.1.0
numpy==1.26.2

# Utilities
python-dotenv==1.0.0
PyYAML==6.0.1

# Logging
loguru==0.7.2
